+++ extras/AFC_HTLF.py
@@ -7,6 +7,9 @@ import traceback
 
 from configparser import Error as error
 
+from extras.AFC_lane import AFCLaneState
+
+
 from extras.AFC_utils import ERROR_STR
 
 try: from extras.AFC_BoxTurtle import afcBoxTurtle
@@ -167,7 +170,7 @@ class AFC_HTLF(afcBoxTurtle):
 
         :return boolean: Returns true if current lane is loaded and printer is printing but lanes status is not ejecting or calibrating
         """
-        return cur_lane.name == self.afc.function.get_current_lane() and self.afc.function.is_printing() and self.status != 'ejecting' and cur_lane.status != "calibrating"
+        return cur_lane.name == self.afc.function.get_current_lane() and self.afc.function.is_printing() and cur_lane.status != AFCLaneState.EJECTING and cur_lane.status != AFCLaneState.CALIBRATING
 
 def load_config_prefix(config):
     return AFC_HTLF(config)