+++ include/update_functions.sh
@@ -31,9 +31,22 @@ AFC Macros update failed.
   fi
   link_extensions
   remove_t_macros
+  remove_velocity
   update_message+="""
 AFC Klipper Add-On updated successfully.
 """
   export update_message
   files_updated_or_installed="True"
+}
+
+remove_velocity() {
+  local files config_file
+  files=$(grep -rlP '^\[AFC_buffer ' "${afc_config_dir}"/*.cfg)
+  for config_file in $files; do
+    section=$(grep -oP '^\[AFC_buffer \K[^\]]+' "$config_file")
+    crudini --del "$config_file" "AFC_buffer $section" velocity
+    export update_message+="""
+Removed deprecated velocity setting from $config_file.
+    """
+  done
 }